openactivedj/openactive/requirements.txt
django-bulk-sync
django-ckeditor